backend/requirements.txt
# ── Core Django ────────────────────────────────────────────────────────────
Django==5.0.6
django-environ==0.11.2

# ── Django REST Framework ──────────────────────────────────────────────────
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2

# ── CORS ───────────────────────────────────────────────────────────────────
django-cors-headers==4.4.0

# ── Database ───────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Production Server ──────────────────────────────────────────────────────
gunicorn==22.0.0

# ── Utilities ──────────────────────────────────────────────────────────────
python-decouple==3.8
